backend/requirements.txt
django==4.2.14
djangorestframework==3.14.0
mongoengine==0.29.1
django-rest-framework-mongoengine==3.4.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
celery==5.4.0
redis==5.0.4
python-dotenv==1.0.1
dnspython==2.6.1
gunicorn==21.2.0
django-cors-headers==4.4.0
pymongo<5          # (optionnel, tu as 4.14.0)

# Modules manquants pour le traitement de fichiers
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
Pillow==10.1.0

# IA et analyse de texte (versions CPU-only plus légères)
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# Certifi pour MongoDB TLS
certifi==2023.11.17

